New syscalls

+++ src/string/strlen.c
@@ -1,7 +1,7 @@
-size_t my_strlen(const unsigned char *str) {
+int my_strlen(const unsigned char *str) {
   const unsigned char *s = str;
   
   while(*s) s++;
 
-  return (size_t)(s - str);
+  return (int)(s - str);
 }
